# docs/api-swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "星云盘 V2 API",
    "description": "星云盘 V2 是一个基于 Go + GORM + Redis + MinIO 的高性能云盘系统。\n\n## 功能特性\n- 用户注册与登录（JWT 认证）\n- 文件上传与下载（支持秒传）\n- 文件夹管理（创建、重命名、移动、删除）\n- 文件去重（基于 SHA256 哈希）\n- 引用计数管理\n\n## 认证方式\n大部分 API 需要在请求头中携带 JWT Token：\n```\nAuthorization: Bearer <access_token>\n```",
    "version": "2.0.0",
    "contact": {
      "name": "星云盘团队",
      "email": "support@xingyunpan.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.xingyunpan.com",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/user/info": {
      "get": {
        "tags": ["用户"],
        "summary": "获取用户信息",
        "description": "获取当前登录用户的详细信息",
        "operationId": "getUserInfo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "example": {
                  "code": 200,
                  "message": "success",
                  "data": {
                    "id": 1,
                    "username": "testuser",
                    "email": "test@example.com",
                    "capacity": 10737418240,
                    "used_size": 1048576,
                    "created_at": "2024-01-01T00:00:00Z",
                    "updated_at": "2024-01-01T00:00:00Z"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}